scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.1",
    "python-dotenv>=1.2.1",
    "web3>=7.14.0",
]
